AMP-28098 using PVs for value validation based on PVs source

+++ amp/WEB-INF/src/org/digijava/kernel/ampapi/endpoints/activity/AmpPossibleValuesDAO.java
@@ -1,5 +1,6 @@
 package org.digijava.kernel.ampapi.endpoints.activity;
 
+import java.math.BigInteger;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -110,8 +111,8 @@ public class AmpPossibleValuesDAO implements PossibleValuesDAO {
     private boolean isValid(String tableName, String configColName, String configType, String idColName, Long id) {
         String query = "SELECT count(" + idColName + ") FROM " + tableName + " WHERE " + configColName + "='"
                 + configType + "' AND " + idColName + "=" + id;
-        Long count = (Long) PersistenceManager.getSession().createSQLQuery(query).uniqueResult();
-        return count == 1;
+        BigInteger count = (BigInteger) PersistenceManager.getSession().createSQLQuery(query).uniqueResult();
+        return count.intValue() == 1;
     }
 
     @Override
